+++ Moco/Tests/testMocoGoals.cpp
@@ -667,3 +667,57 @@ TEMPLATE_TEST_CASE("Endpoint constraint with integral", "", MocoCasADiSolver) {
     // over the motion must be 0.
     CHECK(solution.getControlsTrajectory().norm() < 1e-3);
 }
+
+class MySumSquaredControls : public ModelComponent {
+    OpenSim_DECLARE_CONCRETE_OBJECT(MySumSquaredControls, ModelComponent);
+public:
+    OpenSim_DECLARE_OUTPUT(sum_squared_controls, double,
+            calcSumSquaredControls, SimTK::Stage::Velocity);
+    double calcSumSquaredControls(const SimTK::State& state) const {
+        getModel().realizeVelocity(state);
+        return getModel().getControls(state).normSqr();
+    }
+};
+TEST_CASE("MocoOutputGoal") {
+    auto createStudy = []() {
+        MocoStudy study;
+        study.setName("sliding_mass");
+        MocoProblem& problem = study.updProblem();
+        problem.setModel(createSlidingMassModel());
+        problem.setTimeBounds(0, {0, 5});
+        problem.setStateInfo("/slider/position/value", {0, 1}, 0, 1);
+        problem.setStateInfo("/slider/position/speed", {-100, 100}, 0, 0);
+        problem.setControlInfo("/actuator", MocoBounds(-10, 10));
+        return study;
+    };
+
+    MocoSolution solutionControl;
+    {
+        auto study = createStudy();
+        auto& problem = study.updProblem();
+        problem.addGoal<MocoControlGoal>();
+        auto& solver = study.initCasADiSolver();
+        solver.set_num_mesh_intervals(10);
+        solutionControl = study.solve();
+    }
+    MocoSolution solutionOutput;
+    {
+        auto study = createStudy();
+        auto& problem = study.updProblem();
+        auto model = createSlidingMassModel();
+
+        auto* component = new MySumSquaredControls();
+        component->setName("mysumsquaredcontrols");
+        model->addComponent(component);
+        problem.setModel(std::move(model));
+
+        auto* goal = problem.addGoal<MocoOutputGoal>();
+        goal->setOutputPath("/mysumsquaredcontrols|sum_squared_controls");
+
+        auto& solver = study.initCasADiSolver();
+        solver.set_num_mesh_intervals(10);
+        solutionOutput = study.solve();
+    }
+
+    CHECK(solutionControl.isNumericallyEqual(solutionOutput, 1e-5));
+}