+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- 2019-12-11: Add MocoOutputGoal, allowing any scalar model output to be used as
+              as a goal.
 
 - 2019-12-10: Add an example for MocoInverse in MATLAB, Python, and C++.